src/test/java/inflearn/spring_core/order/OrderServiceImplTest.java
package inflearn.spring_core.order;

import inflearn.spring_core.discount.FixDiscountPolicy;
import inflearn.spring_core.member.Grade;
import inflearn.spring_core.member.Member;
import inflearn.spring_core.member.MemoryMemberRepository;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.*;

class OrderServiceImplTest {

    @DisplayName("주문 생성 테스트")
    @Test
    void createOrder() {
        // 값 다 넣어줌
        MemoryMemberRepository memberRepository = new MemoryMemberRepository();
        memberRepository.save(new Member(1L, "name", Grade.VIP));

        OrderServiceImpl orderService = new OrderServiceImpl(memberRepository, new FixDiscountPolicy());
        Order order = orderService.createOrder(1L, "itemA", 10000);
        assertThat(order.getDiscountPrice()).isEqualTo(1000); // 10퍼센트 할인 - 일단 이것만 확인
    }

}